refactor(remote): extract shared HTTP response handling into RemoteHttpResponseHelper
- Add RemoteHttpResponseHelper with MapToErrorAsync and IsRestrictedHeader
- Move Aether error mapping and status-code-to-Error logic to shared helper
- Consolidate ServiceErrorResponse/ServiceErrorInfo DTOs in helper
- Refactor RemoteInstanceCommandAppService, RemoteInstanceQueryAppService,
  RemoteInstanceRetryAppService, and RemoteAuthorizeAppService to use helper
- Preserve service-specific success handling (e.g. Authorize 200/403, Query 304)

Made-with: Cursor
