+++ src/BBT.Workflow.Infrastructure/Authorization/Remote/RemoteAuthorizeAppService.cs
@@ -86,7 +86,7 @@ public sealed class RemoteAuthorizeAppService(
                 return Result<AuthorizeOutput>.Ok(output ?? new AuthorizeOutput { Allowed = false });
             }
 
-            var error = await MapStatusCodeToErrorCore(response, cancellationToken);
+            var error = await RemoteHttpResponseHelper.MapToErrorAsync(response, cancellationToken, JsonOptions);
             return Result<AuthorizeOutput>.Fail(error);
         }
         catch (Exception ex) when (ex is HttpRequestException or TaskCanceledException or OperationCanceledException)
@@ -128,7 +128,7 @@ public sealed class RemoteAuthorizeAppService(
                 return Result<AuthorizationMatrixOutput>.Ok(output!);
             }
 
-            var error = await MapStatusCodeToErrorCore(response, cancellationToken);
+            var error = await RemoteHttpResponseHelper.MapToErrorAsync(response, cancellationToken, JsonOptions);
             return Result<AuthorizationMatrixOutput>.Fail(error);
         }
         catch (Exception ex) when (ex is HttpRequestException or TaskCanceledException or OperationCanceledException)
@@ -137,106 +137,4 @@ public sealed class RemoteAuthorizeAppService(
         }
     }
 
-    /// <summary>
-    /// Maps HTTP status codes to appropriate Error types following Railway Pattern guidelines.
-    /// - 400 Bad Request → Validation Error
-    /// - 404 Not Found → NotFound Error
-    /// - 409 Conflict → Conflict Error
-    /// - 401 Unauthorized → Unauthorized Error
-    /// - 403 Forbidden → Forbidden Error
-    /// - 5xx Server Error → Dependency Error (external service failed)
-    /// - Other → Dependency Error
-    /// </summary>
-    private static async Task<Error> MapStatusCodeToErrorCore(HttpResponseMessage response,
-        CancellationToken cancellationToken)
-    {
-        var errorContent = await response.ReadDecompressedContentAsync(cancellationToken);
-        var statusCode = response.StatusCode;
-
-        // Check if response has Aether error format header
-        if (response.Headers.TryGetValues("_aether_error_format", out var values) &&
-            values.Any(v => v.Equals("true", StringComparison.OrdinalIgnoreCase)))
-        {
-            try
-            {
-                var errorResponse = JsonSerializer.Deserialize<ServiceErrorResponse>(errorContent, JsonOptions);
-                if (errorResponse?.Error != null)
-                {
-                    var prefix = errorResponse.Error.Prefix ?? InferPrefixFromStatusCode(statusCode);
-                    var code = errorResponse.Error.Code ?? "remote_error";
-
-                    return prefix switch
-                    {
-                        ErrorCodes.Prefixes.Validation => Error.Validation(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.NotFound => Error.NotFound(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Conflict => Error.Conflict(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Unauthorized => Error.Unauthorized(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Forbidden => Error.Forbidden(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Dependency => Error.Dependency(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Transient => Error.Transient(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        _ => Error.Failure(code, errorResponse.Error.Message, errorResponse.Error.Details)
-                    };
-                }
-            }
-            catch (JsonException)
-            {
-                // If JSON deserialization fails, fall back to status code mapping
-            }
-        }
-
-        return statusCode switch
-        {
-            HttpStatusCode.BadRequest => Error.Validation(
-                "remote_bad_request",
-                $"Remote API validation error: {errorContent}"),
-
-            HttpStatusCode.NotFound => Error.NotFound(
-                "remote_not_found",
-                "Requested resource not found on remote API",
-                errorContent),
-
-            HttpStatusCode.Conflict => Error.Conflict(
-                "remote_conflict",
-                $"Remote API conflict: {errorContent}"),
-
-            HttpStatusCode.Unauthorized => Error.Unauthorized(
-                "remote_unauthorized",
-                "Unauthorized access to remote API"),
-
-            HttpStatusCode.Forbidden => Error.Forbidden(
-                "remote_forbidden",
-                "Forbidden access to remote API"),
-
-            HttpStatusCode.InternalServerError or
-                HttpStatusCode.BadGateway or
-                HttpStatusCode.ServiceUnavailable or
-                HttpStatusCode.GatewayTimeout => Error.Dependency(
-                "remote_service_error",
-                $"Remote API service error: {response.ReasonPhrase}",
-                ((int)statusCode).ToString()),
-
-            _ => Error.Dependency(
-                "remote_http_error",
-                $"Remote API returned HTTP {(int)statusCode}: {response.ReasonPhrase}",
-                ((int)statusCode).ToString())
-        };
-    }
-
-    private static string InferPrefixFromStatusCode(HttpStatusCode statusCode)
-    {
-        return statusCode switch
-        {
-            HttpStatusCode.BadRequest => ErrorCodes.Prefixes.Validation,
-            HttpStatusCode.NotFound => ErrorCodes.Prefixes.NotFound,
-            HttpStatusCode.Conflict => ErrorCodes.Prefixes.Conflict,
-            HttpStatusCode.Unauthorized => ErrorCodes.Prefixes.Unauthorized,
-            HttpStatusCode.Forbidden => ErrorCodes.Prefixes.Forbidden,
-            _ => ErrorCodes.Prefixes.Dependency
-        };
-    }
 }
