+++ src/BBT.Workflow.Infrastructure/Instances/Remote/RemoteInstanceQueryAppService.cs
@@ -11,26 +11,6 @@ using Microsoft.Extensions.Options;
 namespace BBT.Workflow.Instances.Remote;
 
 /// <summary>
-/// DTO for Aether error format from remote API responses
-/// </summary>
-internal sealed record ServiceErrorResponse
-{
-    public ServiceErrorInfo? Error { get; init; }
-}
-
-/// <summary>
-/// Error information from Aether format
-/// </summary>
-internal sealed record ServiceErrorInfo
-{
-    public string? Prefix { get; init; }
-    public string? Code { get; init; }
-    public string? Message { get; init; }
-    public string? Details { get; init; }
-    public string? Target { get; init; }
-}
-
-/// <summary>
 /// Remote implementation of instance query operations using HTTP client calls to InstanceController.
 /// Uses IDomainDiscoveryResolver to dynamically resolve endpoint URLs based on target domain.
 /// </summary>
@@ -545,8 +525,8 @@ public sealed class RemoteInstanceQueryAppService(
             return Result<T>.Ok(result!);
         }
 
-        // Map status codes to appropriate Error types (per Railway Pattern)
-        return await MapStatusCodeToError<T>(response, cancellationToken);
+        var error = await RemoteHttpResponseHelper.MapToErrorAsync(response, cancellationToken, JsonOptions);
+        return Result<T>.Fail(error);
     }
 
     /// <summary>
@@ -554,7 +534,6 @@ public sealed class RemoteInstanceQueryAppService(
     /// </summary>
     private static async Task<ConditionalResult<T>> HandleConditionalResponseAsync<T>(HttpResponseMessage response, CancellationToken cancellationToken)
     {
-        // Success case
         if (response.IsSuccessStatusCode)
         {
             var responseContent = await response.ReadDecompressedContentAsync(cancellationToken);
@@ -562,117 +541,7 @@ public sealed class RemoteInstanceQueryAppService(
             return ConditionalResult<T>.Success(result!);
         }
 
-        // Map status codes to appropriate Error types
-        var error = await MapStatusCodeToErrorCore(response, cancellationToken);
+        var error = await RemoteHttpResponseHelper.MapToErrorAsync(response, cancellationToken, JsonOptions);
         return ConditionalResult<T>.Fail(error);
-    }
-
-    /// <summary>
-    /// Maps HTTP status codes to appropriate Error types following Railway Pattern guidelines.
-    /// - 400 Bad Request → Validation Error
-    /// - 404 Not Found → NotFound Error
-    /// - 409 Conflict → Conflict Error
-    /// - 401 Unauthorized → Unauthorized Error
-    /// - 403 Forbidden → Forbidden Error
-    /// - 5xx Server Error → Dependency Error (external service failed)
-    /// - Other → Dependency Error
-    /// </summary>
-    private static async Task<Result<T>> MapStatusCodeToError<T>(HttpResponseMessage response, CancellationToken cancellationToken)
-    {
-        var error = await MapStatusCodeToErrorCore(response, cancellationToken);
-        return Result<T>.Fail(error);
-    }
-
-    private static async Task<Error> MapStatusCodeToErrorCore(HttpResponseMessage response, CancellationToken cancellationToken)
-    {
-        var errorContent = await response.ReadDecompressedContentAsync(cancellationToken);
-        var statusCode = response.StatusCode;
-
-        // Check if response has Aether error format header
-        if (response.Headers.TryGetValues("_aether_error_format", out var values) &&
-            values.Any(v => v.Equals("true", StringComparison.OrdinalIgnoreCase)))
-        {
-            try
-            {
-                var errorResponse = JsonSerializer.Deserialize<ServiceErrorResponse>(errorContent, JsonOptions);
-                if (errorResponse?.Error != null)
-                {
-                    // Use the prefix from remote service if available, otherwise infer from status code
-                    var prefix = errorResponse.Error.Prefix ?? InferPrefixFromStatusCode(statusCode);
-                    var code = errorResponse.Error.Code ?? "remote_error";
-                    
-                    // Map to appropriate Error type based on prefix
-                    return prefix switch
-                    {
-                        ErrorCodes.Prefixes.Validation => Error.Validation(code, errorResponse.Error.Message, errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.NotFound => Error.NotFound(code, errorResponse.Error.Message, errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Conflict => Error.Conflict(code, errorResponse.Error.Message, errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Unauthorized => Error.Unauthorized(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Forbidden => Error.Forbidden(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Dependency => Error.Dependency(code, errorResponse.Error.Message, errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Transient => Error.Transient(code, errorResponse.Error.Message, errorResponse.Error.Target),
-                        _ => Error.Failure(code, errorResponse.Error.Message, errorResponse.Error.Details)
-                    };
-                }
-            }
-            catch (JsonException)
-            {
-                // If JSON deserialization fails, fall back to status code mapping
-            }
-        }
-
-        // Map status code to appropriate Error type (Railway Pattern best practice)
-        return statusCode switch
-        {
-            System.Net.HttpStatusCode.BadRequest => Error.Validation(
-                "remote_bad_request",
-                $"Remote API validation error: {errorContent}"),
-
-            System.Net.HttpStatusCode.NotFound => Error.NotFound(
-                "remote_not_found",
-                "Requested resource not found on remote API",
-                errorContent),
-
-            System.Net.HttpStatusCode.Conflict => Error.Conflict(
-                "remote_conflict",
-                $"Remote API conflict: {errorContent}"),
-
-            System.Net.HttpStatusCode.Unauthorized => Error.Unauthorized(
-                "remote_unauthorized",
-                "Unauthorized access to remote API"),
-
-            System.Net.HttpStatusCode.Forbidden => Error.Forbidden(
-                "remote_forbidden",
-                "Forbidden access to remote API"),
-
-            System.Net.HttpStatusCode.InternalServerError or
-            System.Net.HttpStatusCode.BadGateway or
-            System.Net.HttpStatusCode.ServiceUnavailable or
-            System.Net.HttpStatusCode.GatewayTimeout => Error.Dependency(
-                "remote_service_error",
-                $"Remote API service error: {response.ReasonPhrase}",
-                ((int)statusCode).ToString()),
-
-            _ => Error.Dependency(
-                "remote_http_error",
-                $"Remote API returned HTTP {(int)statusCode}: {response.ReasonPhrase}",
-                ((int)statusCode).ToString())
-        };
-    }
-
-    /// <summary>
-    /// Infers error prefix from HTTP status code for Aether error format
-    /// </summary>
-    private static string InferPrefixFromStatusCode(System.Net.HttpStatusCode statusCode)
-    {
-        return statusCode switch
-        {
-            System.Net.HttpStatusCode.BadRequest => ErrorCodes.Prefixes.Validation,
-            System.Net.HttpStatusCode.NotFound => ErrorCodes.Prefixes.NotFound,
-            System.Net.HttpStatusCode.Conflict => ErrorCodes.Prefixes.Conflict,
-            System.Net.HttpStatusCode.Unauthorized => ErrorCodes.Prefixes.Unauthorized,
-            System.Net.HttpStatusCode.Forbidden => ErrorCodes.Prefixes.Forbidden,
-            _ => ErrorCodes.Prefixes.Dependency
-        };
     }
 }
