+++ src/BBT.Workflow.Infrastructure/Instances/Remote/RemoteInstanceRetryAppService.cs
@@ -79,7 +79,7 @@ public sealed class RemoteInstanceRetryAppService(
             };
 
             var forwardHeaders = currentUser.ToForwardHeaders();
-            CurrentUserForwardHeadersHelper.MergeIntoRequest(requestMessage, forwardHeaders, input.Headers, IsRestrictedHeader);
+            CurrentUserForwardHeadersHelper.MergeIntoRequest(requestMessage, forwardHeaders, input.Headers, RemoteHttpResponseHelper.IsRestrictedHeader);
 
             var response = await httpClient.SendAsync(requestMessage, cancellationToken);
 
@@ -100,7 +100,6 @@ public sealed class RemoteInstanceRetryAppService(
     private static async Task<Result<T>> HandleResponseAsync<T>(HttpResponseMessage response,
         CancellationToken cancellationToken)
     {
-        // Success case
         if (response.IsSuccessStatusCode)
         {
             var responseContent = await response.ReadDecompressedContentAsync(cancellationToken);
@@ -108,143 +107,7 @@ public sealed class RemoteInstanceRetryAppService(
             return Result<T>.Ok(result!);
         }
 
-        // Map status codes to appropriate Error types (per Railway Pattern)
-        return await MapStatusCodeToError<T>(response, cancellationToken);
+        var error = await RemoteHttpResponseHelper.MapToErrorAsync(response, cancellationToken, JsonOptions);
+        return Result<T>.Fail(error);
     }
-
-    /// <summary>
-    /// Maps HTTP status codes to appropriate Error types following Railway Pattern guidelines.
-    /// </summary>
-    private static async Task<Result<T>> MapStatusCodeToError<T>(HttpResponseMessage response,
-        CancellationToken cancellationToken)
-    {
-        var errorContent = await response.ReadDecompressedContentAsync(cancellationToken);
-        var statusCode = response.StatusCode;
-
-        // Check if response has Aether error format header
-        if (response.Headers.TryGetValues("_aether_error_format", out var values) &&
-            values.Any(v => v.Equals("true", StringComparison.OrdinalIgnoreCase)))
-        {
-            try
-            {
-                var errorResponse = JsonSerializer.Deserialize<ServiceErrorResponse>(errorContent, JsonOptions);
-                if (errorResponse?.Error != null)
-                {
-                    // Use the prefix from remote service if available, otherwise infer from status code
-                    var prefix = errorResponse.Error.Prefix ?? InferPrefixFromStatusCode(statusCode);
-                    var code = errorResponse.Error.Code ?? "remote_error";
-
-                    // Map to appropriate Error type based on prefix
-                    var error = prefix switch
-                    {
-                        ErrorCodes.Prefixes.Validation => Error.Validation(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.NotFound => Error.NotFound(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Conflict => Error.Conflict(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Unauthorized => Error.Unauthorized(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Forbidden => Error.Forbidden(code, errorResponse.Error.Message),
-                        ErrorCodes.Prefixes.Dependency => Error.Dependency(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        ErrorCodes.Prefixes.Transient => Error.Transient(code, errorResponse.Error.Message,
-                            errorResponse.Error.Target),
-                        _ => Error.Failure(code, errorResponse.Error.Message, errorResponse.Error.Details)
-                    };
-                    return Result<T>.Fail(error);
-                }
-            }
-            catch (JsonException)
-            {
-                // If JSON deserialization fails, fall back to status code mapping
-            }
-        }
-
-        // Map status code to appropriate Error type (Railway Pattern best practice)
-        var mappedError = statusCode switch
-        {
-            System.Net.HttpStatusCode.BadRequest => Error.Validation(
-                "remote_bad_request",
-                $"Remote API validation error: {errorContent}"),
-
-            System.Net.HttpStatusCode.NotFound => Error.NotFound(
-                "remote_not_found",
-                "Requested resource not found on remote API",
-                errorContent),
-
-            System.Net.HttpStatusCode.Conflict => Error.Conflict(
-                "remote_conflict",
-                $"Remote API conflict: {errorContent}"),
-
-            System.Net.HttpStatusCode.Unauthorized => Error.Unauthorized(
-                "remote_unauthorized",
-                "Unauthorized access to remote API"),
-
-            System.Net.HttpStatusCode.Forbidden => Error.Forbidden(
-                "remote_forbidden",
-                "Forbidden access to remote API"),
-
-            System.Net.HttpStatusCode.InternalServerError or
-                System.Net.HttpStatusCode.BadGateway or
-                System.Net.HttpStatusCode.ServiceUnavailable or
-                System.Net.HttpStatusCode.GatewayTimeout => Error.Dependency(
-                    "remote_service_error",
-                    $"Remote API service error: {response.ReasonPhrase}",
-                    ((int)statusCode).ToString()),
-
-            _ => Error.Dependency(
-                "remote_http_error",
-                $"Remote API returned HTTP {(int)statusCode}: {response.ReasonPhrase}",
-                ((int)statusCode).ToString())
-        };
-
-        return Result<T>.Fail(mappedError);
-    }
-
-    /// <summary>
-    /// Infers error prefix from HTTP status code for Aether error format
-    /// </summary>
-    private static string InferPrefixFromStatusCode(System.Net.HttpStatusCode statusCode)
-    {
-        return statusCode switch
-        {
-            System.Net.HttpStatusCode.BadRequest => ErrorCodes.Prefixes.Validation,
-            System.Net.HttpStatusCode.NotFound => ErrorCodes.Prefixes.NotFound,
-            System.Net.HttpStatusCode.Conflict => ErrorCodes.Prefixes.Conflict,
-            System.Net.HttpStatusCode.Unauthorized => ErrorCodes.Prefixes.Unauthorized,
-            System.Net.HttpStatusCode.Forbidden => ErrorCodes.Prefixes.Forbidden,
-            _ => ErrorCodes.Prefixes.Dependency
-        };
-    }
-
-    /// <summary>
-    /// Checks if a header is restricted and cannot be set manually
-    /// </summary>
-    private static bool IsRestrictedHeader(string headerName)
-    {
-        return headerName.Equals("Content-Type", StringComparison.OrdinalIgnoreCase) ||
-               headerName.Equals("Content-Length", StringComparison.OrdinalIgnoreCase) ||
-               headerName.Equals("Host", StringComparison.OrdinalIgnoreCase) ||
-               headerName.Equals("Connection", StringComparison.OrdinalIgnoreCase);
-    }
-}
-
-/// <summary>
-/// DTO for deserializing error responses from remote services
-/// </summary>
-internal sealed class ServiceErrorResponse
-{
-    public ServiceError? Error { get; set; }
-}
-
-/// <summary>
-/// DTO for service error details
-/// </summary>
-internal sealed class ServiceError
-{
-    public string? Code { get; set; }
-    public string? Message { get; set; }
-    public string? Prefix { get; set; }
-    public string? Target { get; set; }
-    public string? Details { get; set; }
 }
